Now sending the actual photo to the server.
Instead of sending some test text, now sending the bitmap in the form of
a Byte array via POST to a php script. The server is now setup correctly
and we are receiving HTTP code 200, so that's good. Added a CJL (my
initials) tag to the logging so I can filter in logcat, as the actual
device produces a lot of logs from other services.